backend/requirement.txt
Flask
flask-cors
PyPDF2
langchain
langchain-community
sentence-transformers
psycopg2-binary
python-dotenv
google-generativeai